# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: YuTong Platform API
  description: "YuTong é\x9B¨æ¡\x90 - å\x95\x86ä¸\x9Açº§å\N¨æ \x88æ\x8A\x80æ\x9C¯åº\x95\
    åº§ API å¥\x91çº¦"
  contact:
    name: YuTong Team
  version: 0.2.0
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/lowcode/entities/{id}:
    get:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "å®\x9Eä½\x93è¯¦æ\x83\N"
      operationId: getLowcodeEntity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultLcEntityDetailVO'
    put:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "æ\x9B´æ\x96°å®\x9Eä½\x93"
      operationId: updateLowcodeEntity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLcEntityRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultLcEntity'
...
  /api/v1/todos/{id}/complete:
    post:
      tags:
      - "å¾\Nå\x8A\x9Eç®¡ç\x90\x86"
      summary: "å®\x8Cæ\x88\x90å¾\Nå\x8A\x9E"
      operationId: completeTodo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
...
  /api/v1/messages/{id}/read:
    post:
      tags:
      - "ç«\x99å\x86\Nä¿¡ç®¡ç\x90\x86"
      summary: "æ \x87è®°æ¶\x88æ\x81¯å·²è¯»"
      operationId: markMessageAsRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
...
  /api/v1/lowcode/entities:
    get:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢å®\x9Eä½\x93"
      operationId: listLowcodeEntities
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: entityCode
        in: query
        required: false
        schema:
          type: string
      - name: entityName
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultLcEntity'
    post:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "å\x88\x9Bå»ºå®\x9Eä½\x93è\x8D\x89ç¨¿"
      operationId: createLowcodeEntityDraft
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLcEntityRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultLcEntity'
  /api/v1/lowcode/entities/{id}/publish:
    post:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "å\x8F\x91å¸\x83å®\x9Eä½\x93ç\x89\x88æ\x9C¬"
      operationId: publishLowcodeEntity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultLcEntity'
  /api/v1/lowcode/entities/{id}/disable:
    post:
      tags:
      - "ä½\x8Eä»£ç \x81-å®\x9Eä½\x93"
      summary: "ç¦\x81ç\x94¨å®\x9Eä½\x93"
      operationId: disableLowcodeEntity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultLcEntity'
...
  /api/v1/todos:
    get:
      tags:
      - "å¾\Nå\x8A\x9Eç®¡ç\x90\x86"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢å½\x93å\x89\x8Dç\x94¨æ\x88·å¾\Nå\x8A\x9E"
      operationId: pageTodos
      parameters:
      - name: pageNo
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: todoStatus
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultSysTodoTask'
...
  /api/v1/messages:
    get:
      tags:
      - "ç«\x99å\x86\Nä¿¡ç®¡ç\x90\x86"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢å½\x93å\x89\x8Dç\x94¨æ\x88·ç«\x99å\x86\Nä¿¡"
      operationId: pageMessages
      parameters:
      - name: pageNo
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: readStatus
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultSysMessage'
...
components:
  schemas:
...
    LcFieldDTO:
      type: object
      properties:
        id:
          type: string
        fieldCode:
          type: string
        fieldName:
          type: string
        dbColumn:
          type: string
        dataType:
          type: string
        lengthValue:
          type: integer
          format: int32
        precisionValue:
          type: integer
          format: int32
        scaleValue:
          type: integer
          format: int32
        nullable:
          type: boolean
        defaultValue:
          type: string
        dictType:
          type: string
        primaryFlag:
          type: boolean
        uniqueFlag:
          type: boolean
        indexFlag:
          type: boolean
        sortNo:
          type: integer
          format: int32
    LcRelationDTO:
      type: object
      properties:
        targetEntityId:
          type: string
        relationType:
          type: string
        sourceFieldCode:
          type: string
        targetFieldCode:
          type: string
        cascadePolicy:
          type: string
        required:
          type: boolean
    SaveLcEntityRequest:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        entityCode:
          type: string
        entityName:
          type: string
        tableName:
          type: string
        moduleCode:
          type: string
        ownerUserId:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/LcFieldDTO'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/LcRelationDTO'
    LcEntity:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        entityCode:
          type: string
        entityName:
          type: string
        tableName:
          type: string
        moduleCode:
          type: string
        versionNo:
          type: integer
          format: int32
        schemaVersion:
          type: string
        configHash:
          type: string
        status:
          type: string
        ownerUserId:
          type: string
    ResultLcEntity:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/LcEntity'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ResultVoid:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          type: object
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    PageResultSysTodoTask:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysTodoTask'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultSysTodoTask:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultSysTodoTask'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    SysTodoTask:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        todoType:
          type: string
        bizType:
          type: string
        bizId:
          type: string
        title:
          type: string
        assigneeId:
          type: string
        todoStatus:
          type: string
        priority:
          type: string
        dueTime:
          type: string
          format: date-time
        completedTime:
          type: string
          format: date-time
        sourceEventId:
          type: string
...
    PageResultSysMessage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysMessage'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultSysMessage:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultSysMessage'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    SysMessage:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        receiverId:
          type: string
        msgType:
          type: string
        title:
          type: string
        content:
          type: string
        readStatus:
          type: string
        readTime:
          type: string
          format: date-time
        bizType:
          type: string
        bizId:
          type: string
        targetRouteId:
          type: string
        targetParams:
          type: string
...
    PageResultLcEntity:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/LcEntity'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultLcEntity:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultLcEntity'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    LcEntityDetailVO:
      type: object
      properties:
        id:
          type: string
        entityCode:
          type: string
        entityName:
          type: string
        tableName:
          type: string
        moduleCode:
          type: string
        versionNo:
          type: integer
          format: int32
        schemaVersion:
          type: string
        configHash:
          type: string
        status:
          type: string
        ownerUserId:
          type: string
        version:
          type: integer
          format: int32
        createdTime:
          type: string
          format: date-time
        publishedTime:
          type: string
          format: date-time
        fields:
          type: array
          items:
            $ref: '#/components/schemas/LcFieldDTO'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/LcRelationDTO'
    ResultLcEntityDetailVO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/LcEntityDetailVO'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: MOCK